+++ NEWS.md
@@ -1,5 +1,7 @@
 # scribe (development version)
 
+* positional arguments now can have default values [#52](https://github.com/jmbarbone/scribe/issues/52)
+
 # scribe 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.